+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/member/entity/MemberRepository.java
@@ -1,8 +1,6 @@
-package com.NBE_4_5_2.Team5.domain.member.repository;
+package com.NBE_4_5_2.Team5.domain.member.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.NBE_4_5_2.Team5.domain.member.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, String> {
 }
