+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/payment/service/PaymentService.java
@@ -1,31 +1,22 @@
 package com.NBE_4_5_2.Team5.domain.payment.service;
 
-import java.util.Map;
 import java.util.UUID;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.client.RestClient;
 
 import com.NBE_4_5_2.Team5.domain.member.entity.Member;
-import com.NBE_4_5_2.Team5.domain.member.repository.MemberRepository;
+import com.NBE_4_5_2.Team5.domain.member.entity.MemberRepository;
 import com.NBE_4_5_2.Team5.domain.payment.dto.PaymentDto;
 import com.NBE_4_5_2.Team5.domain.payment.dto.PaymentMetaData;
 import com.NBE_4_5_2.Team5.domain.payment.entity.Payment;
 import com.NBE_4_5_2.Team5.domain.payment.enums.PaymentStatus;
 import com.NBE_4_5_2.Team5.domain.payment.repository.PaymentRepository;
 import com.NBE_4_5_2.Team5.domain.product.entity.Product;
-import com.NBE_4_5_2.Team5.domain.product.repository.ProductRepository;
-import com.NBE_4_5_2.Team5.global.exception.TossPaymentException;
+import com.NBE_4_5_2.Team5.domain.product.entity.ProductRepository;
 import com.NBE_4_5_2.Team5.global.exception.product.ProductNotFoundException;
 
 import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -38,9 +29,8 @@ public class PaymentService {
 	private final PaymentRepository paymentRepository;
 	private final MemberRepository memberRepository;
 	private final ProductRepository productRepository;
-
-	@Value("${custom.toss.payment.secret}")
-	private String tossPaymentSecretKey;
+	// TODO : custom annotation으로 다른 PG사의 Adapter bean을 구분해서 가져오게 한다면?
+	private final PaymentProviderAdapter paymentProviderAdapter;
 
 	public PaymentMetaData saveMetaData(String paymentId, Integer amount) {
 		Member user = getUser();
@@ -79,15 +69,6 @@ public class PaymentService {
 		return PaymentDto.of(saved);
 	}
 
-	@Getter
-	@NoArgsConstructor
-	@AllArgsConstructor
-	private static class TossPaymentReqBody {
-		private String paymentKey;
-		private String orderId;
-		private int amount;
-	}
-
 	public void requestCharge(@NotNull String id, @NotNull String paymentKey, @NotNull Integer amount) {
 		Member loggedInUser = getUser();
 		Payment payment = paymentRepository.findById(id)
@@ -100,25 +81,9 @@ public class PaymentService {
 		Payment update = payment.updatePaymentKey(paymentKey);
 
 		try {
-			RestClient restClient = RestClient.create();
-			ResponseEntity<Map> response = restClient.post()
-				.uri("https://api.tosspayments.com/v1/payments/confirm")
-				.header("Content-Type", "application/json")
-				.header("Authorization", "Basic " + tossPaymentSecretKey)
-				.body(new TossPaymentReqBody(update.getPaymentKey(), update.getId(), update.getTotalPrice()))
-				.retrieve()
-				.toEntity(Map.class);
+			paymentProviderAdapter.requestPayment(id, paymentKey, amount);
 
-			if (!response.getStatusCode().equals(HttpStatus.OK) || !response.hasBody()) {
-				throw new TossPaymentException("결제 승인 요청에 실패했습니다.");
-			}
-
-			@SuppressWarnings("unchecked")
-			Map<String, Object> body = (Map<String, Object>)response.getBody();
-			assert body != null;
-			Integer totalAmount = (Integer)body.get("totalAmount");
-
-			loggedInUser.chargeCash(totalAmount);
+			loggedInUser.chargeCash(amount);
 
 			payment.updateState(PaymentStatus.DONE);
 
@@ -126,4 +91,5 @@ public class PaymentService {
 			log.error(e.getMessage());
 		}
 	}
+
 }
