+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/product/entity/ProductRepository.java
@@ -1,8 +1,6 @@
-package com.NBE_4_5_2.Team5.domain.product.repository;
+package com.NBE_4_5_2.Team5.domain.product.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.NBE_4_5_2.Team5.domain.product.entity.Product;
 
 public interface ProductRepository extends JpaRepository<Product, String> {
 }
